+++ api/test/unit/konnect_funcs_test.go
@@ -304,7 +304,7 @@ func TestServiceRef(t *testing.T) {
 
 			serviceRef := &configurationv1alpha1.ServiceRef{
 				Type: configurationv1alpha1.ServiceRefNamespacedRef,
-				NamespacedRef: &commonv1alpha1.NameRef{
+				NamespacedRef: &commonv1alpha1.NamespacedRef{
 					Name: "test-service",
 				},
 			}