+++ controller/konnect/reconciler_serviceref_test.go
@@ -128,7 +128,7 @@ func TestHandleServiceRef(t *testing.T) {
 				Spec: configurationv1alpha1.KongRouteSpec{
 					ServiceRef: &configurationv1alpha1.ServiceRef{
 						Type: configurationv1alpha1.ServiceRefNamespacedRef,
-						NamespacedRef: &commonv1alpha1.NameRef{
+						NamespacedRef: &commonv1alpha1.NamespacedRef{
 							Name: "svc-ok",
 						},
 					},
@@ -166,7 +166,7 @@ func TestHandleServiceRef(t *testing.T) {
 				Spec: configurationv1alpha1.KongRouteSpec{
 					ServiceRef: &configurationv1alpha1.ServiceRef{
 						Type: configurationv1alpha1.ServiceRefNamespacedRef,
-						NamespacedRef: &commonv1alpha1.NameRef{
+						NamespacedRef: &commonv1alpha1.NamespacedRef{
 							Name: "svc-ok",
 						},
 					},
@@ -189,7 +189,7 @@ func TestHandleServiceRef(t *testing.T) {
 				Spec: configurationv1alpha1.KongRouteSpec{
 					ServiceRef: &configurationv1alpha1.ServiceRef{
 						Type: configurationv1alpha1.ServiceRefNamespacedRef,
-						NamespacedRef: &commonv1alpha1.NameRef{
+						NamespacedRef: &commonv1alpha1.NamespacedRef{
 							Name: "svc-not-programmed",
 						},
 					},
@@ -228,7 +228,7 @@ func TestHandleServiceRef(t *testing.T) {
 				Spec: configurationv1alpha1.KongRouteSpec{
 					ServiceRef: &configurationv1alpha1.ServiceRef{
 						Type: configurationv1alpha1.ServiceRefNamespacedRef,
-						NamespacedRef: &commonv1alpha1.NameRef{
+						NamespacedRef: &commonv1alpha1.NamespacedRef{
 							Name: "svc-with-cp-ref-unprogrammed",
 						},
 					},