+++ test/crdsvalidation/configuration.konghq.com/kongroute_test.go
@@ -44,7 +44,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type: configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{
+							NamespacedRef: &commonv1alpha1.NamespacedRef{
 								Name: "test-konnect-service",
 							},
 						},
@@ -66,7 +66,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type: configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{
+							NamespacedRef: &commonv1alpha1.NamespacedRef{
 								Name: "test-konnect-service",
 							},
 						},
@@ -89,7 +89,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type: configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{
+							NamespacedRef: &commonv1alpha1.NamespacedRef{
 								Name: "test-konnect-service",
 							},
 						},
@@ -106,7 +106,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type: configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{
+							NamespacedRef: &commonv1alpha1.NamespacedRef{
 								Name: "",
 							},
 						},
@@ -160,7 +160,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type:          configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{Name: "svc"},
+							NamespacedRef: &commonv1alpha1.NamespacedRef{Name: "svc"},
 						},
 						KongRouteAPISpec: configurationv1alpha1.KongRouteAPISpec{
 							Paths: []string{"/"},
@@ -175,7 +175,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type:          configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{Name: "svc"},
+							NamespacedRef: &commonv1alpha1.NamespacedRef{Name: "svc"},
 						},
 						KongRouteAPISpec: configurationv1alpha1.KongRouteAPISpec{
 							Protocols: []sdkkonnectcomp.RouteJSONProtocols{"http"},
@@ -191,7 +191,7 @@ func TestKongRoute(t *testing.T) {
 					Spec: configurationv1alpha1.KongRouteSpec{
 						ServiceRef: &configurationv1alpha1.ServiceRef{
 							Type:          configurationv1alpha1.ServiceRefNamespacedRef,
-							NamespacedRef: &commonv1alpha1.NameRef{Name: "svc"},
+							NamespacedRef: &commonv1alpha1.NamespacedRef{Name: "svc"},
 						},
 						KongRouteAPISpec: configurationv1alpha1.KongRouteAPISpec{
 							Protocols: []sdkkonnectcomp.RouteJSONProtocols{"http"},