+++ CHANGELOG.md
@@ -47,6 +47,18 @@
   [#3015](https://github.com/Kong/kong-operator/issues/3015)
 - HybridGateway: Added support to PathPrefixMatch for the `URLRewrite` `HTTPRoute` filter.
   [#3039](https://github.com/Kong/kong-operator/pull/3039)
+- HybridGateway: Added comprehensive HTTPRoute converter tests to improve translation stability.
+  [#3111](https://github.com/Kong/kong-operator/pull/3111)
+- Support cross namespace references from `KongPluginBinding` to `KongPlugin`.
+  For this reference to be allowed, a `KongReferenceGrant` resource must be created
+  in the namespace of the `KongPlugin`, allowing access for the `KongPluginBinding`.
+  [#3108](https://github.com/Kong/kong-operator/pull/3108)
+- HybridGateway: Added support to PathPrefixMatch for the `RequestRedirect` `HTTPRoute` filter.
+  [#3065](https://github.com/Kong/kong-operator/pull/3065)
+- Support cross namespace references from `KongRoute` to `KongService`.
+  For this reference to be allowed, a `KongReferenceGrant` resource must be created
+  in the namespace of the `KongService`, allowing access for the `KongRoute`.
+  [#3125](https://github.com/Kong/kong-operator/pull/3125)
 
 ### Fixes